+++ src/Amadeus/Client/Struct/Pnr/NameChange/DateOfBirthInEnhancedPaxData.php
@@ -20,14 +20,28 @@
  * @license https://opensource.org/licenses/Apache-2.0 Apache 2.0
  */
 
-namespace Amadeus\Client\Struct;
+namespace Amadeus\Client\Struct\Pnr\NameChange;
 
 /**
- * BaseWsMessage Base Web Service message class
+ * DateOfBirthInEnhancedPaxData
  *
- * @package Amadeus\Client\Struct
+ * @package Amadeus\Client\Struct\Pnr\NameChange
  * @author Dieter Devlieghere <dieter.devlieghere@benelux.amadeus.com>
  */
-class BaseWsMessage extends WsMessageUtility
+class DateOfBirthInEnhancedPaxData
 {
+    /**
+     * @var DateTime
+     */
+    public $dateTime;
+
+    /**
+     * DateOfBirthInEnhancedPaxData constructor.
+     *
+     * @param \DateTime $dateOfBirth
+     */
+    public function __construct($dateOfBirth)
+    {
+        $this->dateTime = new DateTime($dateOfBirth);
+    }
 }