+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release 1.2.0 (Unreleased)
 * Request Creator split up in 1 dedicated class per message.
-* Implemented ``Ticket_DeleteTSMP``
 * Added support for requesting rules for specific Fare Components after a pricing request in ``Fare_CheckRules`` (https://github.com/amabnl/amadeus-ws-client/issues/21)
+* Implemented ``Ticket_DeleteTSMP`` message
+* Implemented ``PNR_NameChange`` message
 
 # Release 1.1.0 (19 October 2016)